+++ python/pycrdt/_array.py
@@ -213,20 +213,29 @@ class Array(BaseType, Generic[T]):
         Raises:
             RuntimeError: Index must be of type integer.
         """
-        with self.doc.transaction():
+        with self.doc.transaction() as txn:
             if isinstance(key, int):
                 key = self._check_index(key)
                 del self[key]
                 self[key:key] = [value]
             elif isinstance(key, slice):
+                end = len(self)
+                start = key.start if key.start is not None else 0
+                stop = key.stop if key.stop is not None else end
                 if key.step is not None:
                     raise RuntimeError("Step not supported")
-                if key.start != key.stop:
-                    raise RuntimeError("Start and stop must be equal")
-                if key.start > len(self) or key.start < 0:
+                if start > len(self) or start < 0:
                     raise RuntimeError("Index out of range")
-                for i, v in enumerate(value):
-                    self._set(i + key.start, v)
+                iter_value = iter(value)
+                # remove existing elements
+                self.integrated.remove_range(
+                    txn._txn,
+                    start,
+                    stop - start,
+                )
+                # replace existing elements with new ones
+                for i, v in enumerate(iter_value):
+                    self._set(i + start, v)
             else:
                 raise RuntimeError("Index must be of type integer")